feat: move scroll events
The scroll events are now handled by Stimulus as well.
Now, all JavaScript is embedded within Stimulus.
