+++ src/sphinxawesome_theme/layout.html
@@ -92,7 +92,11 @@
         </aside>
       {%- endif %}
 
-      <main class="px-5 lg:ml-fluid grid-area-main overflow-y-auto flex flex-col h-full {% if theme_show_nav|lower != 'true' -%}mx-auto{%- endif -%}">
+      <main
+        data-controller="scroll"
+        data-action="scroll->scroll#showButton"
+        class="px-5 lg:ml-fluid grid-area-main overflow-y-auto flex flex-col h-full {% if theme_show_nav|lower != 'true' -%}mx-auto{%- endif -%}"
+      >
         {%- if pagename != master_doc and theme_show_breadcrumbs|lower == 'true' %}
           {%- include "breadcrumbs.html" %}
         {% endif -%}
@@ -115,6 +119,25 @@
             {% include "footer.html" %}
           {% endblock %}
         </footer>
+
+        {#- scroll to top button -#}
+        {%- if theme_show_scrolltop|lower == "true" -%}
+        <button
+          id="scrolltop"
+          data-scroll-target="scrollToTop"
+          data-action="scroll#scrollToTop"
+          class="fixed bottom-0 right-0 p-2 m-4 rounded-full bg-gray-700 text-white opacity-0 transition-opacity duration-1000 pointer-events-none hover:bg-gray-200 hover:text-pink-500 focus:outline-none focus:ring-2 focus:ring-blue-600 focus:bg-gray-200 focus:text-pink-500 tooltipped tooltipped-nw"
+          aria-label="{{ _('Scroll to top') }}"
+        >
+          <svg
+            class="fill-current stroke-current h-8"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 24 24"
+          >
+            <path d="M7.41 15.41L12 10.83l4.59 4.58L18 14l-6-6-6 6z" />
+          </svg>
+        </button>
+        {%- endif -%}
       </main>
 
       {% include "menu.html" %}
